+++ software/firmware/europi_script.py
@@ -1,29 +1,43 @@
 import os
 
 class EuroPiScript:
-    """Base class for scripts which wish to participate in the bootloader menu.
+    """A Base class for scripts which wish to participate in the bootloader menu.
      
-    To make your script compatible with the menu, you must define your script
-    as a class that inherits this base class. For example::
-        from europi_script import EuroPiScript
-        
-        MyScript(EuroPiScript):
-            # init and other methods
-            # Override main with your script main method.
-            def main(self):
-                # Your script's main loop.
+    To make your script compatible with the menu, you must:
+      
+      * define your script as a class that inherits from this base class.
+      * override ``main()`` to define your script's main loop
+
+    For example::
+    
+       from europi_script import EuroPiScript
+    
+       class HelloWorld(EuroPiScript):
+           # init and other methods
+           # Override main with your script main method.
+           def main(self):
+               # Your script's main loop.
     
     """
 
     def __init__(self) -> None:
         self._state_filename = f"saved_state_{self.display_name()}.txt"
 
-    def main(self):
+    def main(self) -> None:
         """Override this method with your script's main loop method."""
         raise NotImplementedError
 
     @classmethod
-    def display_name(cls):
+    def display_name(cls) -> str:
+        """Returns the string used to identify this script in the Menu. Defaults to the class name. Override it if you 
+        would like to use a different name::
+        
+           @classmethod
+           def display_name(cls):
+               return "Hello World"
+
+        Note that the screen is only 16 characters wide. Anything longer will be cut off.
+        """
         return cls.__qualname__
 
     #